refactor : 카멜 케이스로 네이밍 수정 [#09]
