+++ frontend/src/features/chat/provider/ChatProvider.tsx
@@ -1,5 +1,5 @@
-import { useSocketIO } from "@/shared/hooks/use-socket-io";
-import { useUserContext } from "@/shared/hooks/use-user-context";
+import { useSocketIO } from "@/shared/hooks/useSocketIo";
+import { useUserContext } from "@/shared/hooks/useUserContext";
 import React from "react";
 
 interface ChatContextType {
