+++ frontend/src/features/waiting-room/hooks/use-waiting-context.tsx
