+++ frontend/src/features/waiting-room/provider/WaitingRoomProvider.tsx
@@ -1,4 +1,4 @@
-import { useSocketIO } from "@/shared/hooks/use-socket-io";
+import { useSocketIO } from "@/shared/hooks/useSocketIo";
 import { responseBetRoomInfo } from "@betting-duck/shared";
 import { useLoaderData, useLocation, useRouter } from "@tanstack/react-router";
 import React from "react";
