--- frontend/src/routes/betting-page.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { BettingPage } from "@/features/betting-page";
-
-export const Route = createFileRoute("/betting-page")({
-  component: BettingPage,
-});
