+++ frontend/src/routes/betting.index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { WaitingError } from "@/features/waiting-room/ui/WaitingError";
-import { useUserContext } from "@/shared/hooks/use-user-context";
+import { useUserContext } from "@/shared/hooks/useUserContext";
 import { createFileRoute, redirect, useNavigate } from "@tanstack/react-router";
 import { ROUTES } from "@/shared/config/route";
 import { responseBetRoomInfo } from "@betting-duck/shared";
